# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Display a user listing.",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Users index."
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add a new user.",
                "operationId": "create",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "default": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by id.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the user that needs to be fetched.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update an existing user.",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the user that needs to be fetched.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deletes a user",
                "description": "",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id to delete.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    }
}